requirements.txt
Flask==3.0.2
joblib==1.3.2
numpy==1.24.4
pandas==1.5.3
Pillow==10.3.0
tensorflow==2.16.1
gunicorn==21.2.0
